requirements.txt
# Core Django
Django>=6.0.0
django-cors-headers>=4.3.0

# AI & API
openai>=1.17.1
httpx>=0.27.0

# Environment & Configuration
python-dotenv>=1.0.1

# Production Server (for deployment)
gunicorn>=21.2.0
whitenoise>=6.6.0

# Database (optional - for PostgreSQL/MySQL)
# psycopg2-binary>=2.9.9  # PostgreSQL adapter
# mysqlclient>=2.2.0  # MySQL adapter

# Image Processing (if needed for future features)
# Pillow>=10.2.0

# Security & Rate Limiting
django-ratelimit>=4.1.0  # Rate limiting for API endpoints
djangorestframework>=3.15.2

# Logging & Monitoring
# sentry-sdk>=2.0.0  # Error tracking (optional)

# Development Tools (optional)
# ipython>=8.20.0  # Enhanced Python shell
# django-extensions>=3.2.3  # Useful Django management commands
# django-debug-toolbar>=4.2.0  # Debug toolbar for development

# Testing (optional)
# pytest>=7.4.0
# pytest-django>=4.7.0
# coverage>=7.3.0  # Code coverage

# Code Quality (optional)
# black>=23.12.0  # Code formatter
# flake8>=6.1.0  # Linter
# isort>=5.13.0  # Import sorter

# Optional additions
streamlit>=1.35.0
